third_party/WebKit/LayoutTests/bluetooth/connectGATT.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/bluetooth-helpers.js"></script>
<script>
'use strict';

test(t => { assert_true(window.testRunner instanceof Object); t.done(); },
     'window.testRunner is required for the following tests.');

promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => {
      testRunner.setBluetoothMockDataSet('EmptyAdapter');
      return assert_promise_rejects_with_message(
        device.connectGATT(),
        new DOMException('Bluetooth Device is no longer in range.',
                         'NetworkError'),
        'Device went out of range.');
    });
}, 'Device goes out of range. Reject with NetworkError.');

// The following tests make sure the Web Bluetooth implementation
// responds correctly to the different types of errors the
// underlying platform might throw.

// Each implementation maps these devices to specific code paths
// that result in different errors thus increasing code coverage
// when testing. Therefore some of these devices might not be useful
// for all implementations.
[{
  testName: 'Unknown error when connnecting.',
  uuid: errorUUID(0x0),
  error: new DOMException('Unknown error when connecting to the device.',
                          'NetworkError')
}, {
  testName: 'Connection was already in progress.',
  uuid: errorUUID(0x1),
  error: new DOMException('Connection already in progress.',
                          'NetworkError')
}, {
  testName: 'Connection failed.',
  uuid: errorUUID(0x2),
  error: new DOMException('Connection failed for unknown reason.',
                          'NetworkError')
}, {
  testName: 'Authentication failed when connecting.',
  uuid: errorUUID(0x3),
  error: new DOMException('Authentication failed.',
                          'NetworkError')
}, {
  testName: 'Authentication canceled when connecting.',
  uuid: errorUUID(0x4),
  error: new DOMException('Authentication canceled.',
                          'NetworkError')
}, {
  testName: 'Authentication rejected when connecting.',
  uuid: errorUUID(0x5),
  error: new DOMException('Authentication rejected.',
                          'NetworkError')
}, {
  testName: 'Authentication timed out when connecting.',
  uuid: errorUUID(0x6),
  error: new DOMException('Authentication timeout.',
                          'NetworkError')
}, {
  testName: 'Tried to connect to an unsupported device.',
  uuid: errorUUID(0x7),
  error: new DOMException('Unsupported device.',
                          'NetworkError')
}, {
  testName: 'A write operation exceeds the maximum length of the attribute.',
  uuid: errorUUID(0x8),
  error: new DOMException('Attribute length invalid.',
                          'NetworkError')
}, {
  testName: 'A remote device connection is congested.',
  uuid: errorUUID(0x9),
  error: new DOMException('Connection congested.',
                          'NetworkError')
}, {
  testName: 'Insufficient encryption for a given operation.',
  uuid: errorUUID(0xa),
  error: new DOMException('Insufficient encryption.',
                          'NetworkError')
}, {
  testName: 'A read or write operation was requested with an invalid offset.',
  uuid: errorUUID(0xb),
  error: new DOMException('Offset invalid.',
                          'NetworkError')
}, {
  testName: 'GATT read operation is not permitted.',
  uuid: errorUUID(0xc),
  error: new DOMException('Read not permitted.',
                          'NetworkError')
}, {
  testName: 'The given request is not supported.',
  uuid: errorUUID(0xd),
  error: new DOMException('Request not supported.',
                          'NetworkError')
}, {
  testName: 'GATT write operation is not permitted.',
  uuid: errorUUID(0xe),
  error: new DOMException('Write not permitted.',
                          'NetworkError')
}].forEach(testSpec => {
  promise_test(() => {
    testRunner.setBluetoothMockDataSet('FailingConnectionsAdapter');
    return requestDeviceWithKeyDown({filters: [{services: [testSpec.uuid]}]})
      .then(device => {
        assert_promise_rejects_with_message(
          device.connectGATT(),
          testSpec.error,
          'Adapter failed to connect to device.');
      });
  }, testSpec.testName);
});

// TODO(ortuno): Allow connections when the tab is in the background.
// This is a short term solution instead of implementing a tab indicator
// for bluetooth connections.
// https://crbug.com/579746
promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => {
      testRunner.setPageVisibility('hidden');
      return assert_promise_rejects_with_message(
        device.connectGATT(),
        new DOMException('Connection is only allowed while the page is visible. ' +
                         'This is a temporary measure until we are able to ' +
                         'effectively communicate to the user that a page is ' +
                         'connected to a device.',
                         'SecurityError'));
    })
    .then(() => testRunner.setPageVisibility('visible'))
    .catch(error => {
      testRunner.setPageVisibility('visible');
      throw error;
    });
}, 'Device should not be able to connect in background.');

promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => device.connectGATT())
    .then(gattServer => assert_true(gattServer.connected));
}, 'Device will connect');

promise_test(() => {
  testRunner.setBluetoothMockDataSet('');
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => {
      return Promise.all([device.connectGATT(), device.connectGATT()])
    }).then(gattServers => {
      // connectGATT should return the same object if it was created earlier.
      // TODO(ortuno): change to assert_equals.
      assert_not_equals(gattServers[0], gattServers[1]);
    });
});
</script>
